upd5

+++ My_project/Assets/_Project/Classes/NPC/Fraction/Fraction.cs
@@ -9,10 +9,10 @@
         public readonly string Color;
         public readonly EFractionTypes FractionType;
 
-        public float Money;
+        public long Money;
         
         public Fraction(int id, string name, int homeSector = 0, int homeConstellationId = 0, 
-        string color = null, EFractionTypes fractionType = EFractionTypes.Regular, float money = 0)
+        string color = null, EFractionTypes fractionType = EFractionTypes.Regular, long money = 0)
         {
             Id = id;
             Name = name;
