+++ My_project/Assets/_Project/Classes/Trade/Models/ITradeActor.cs
@@ -8,6 +8,6 @@ namespace _Project.Scripts.Trade.Models
     {
         Fraction Owner { get; }
         Cargo Cargo { get; }
-        float Money { get; }
+        long Money { get; }
     }
 }
